stores: add regression test

+++ stores/slabbuffer.go
@@ -306,7 +306,7 @@ func (mgr *SlabBufferManager) FetchPartialSlab(ctx context.Context, ec object.En
 	}
 
 	data := make([]byte, length)
-	_, err := buffer.file.ReadAt(data, offset)
+	_, err := buffer.file.ReadAt(data, int64(offset))
 	if err != nil {
 		return nil, fmt.Errorf("failed to read data from buffer (offset: %v, length: %v): %w", offset, length, err)
 	}
